add get search by three cards
