+++ src/components/App/App.js
@@ -6,48 +6,9 @@ import Footer from '../Footer/Footer';
 import Header from '../Header/Header';
 import PopupWithForm from '../PopupWithForm/PopupWithForm';
 import SavedNews from '../SavedNews/SavedNews';
+import {cards, user} from '../../utils/constants';
 
 function App() {
-
-  const cards = [
-    {
-      keyWord: "Ладога",
-      date: "2 августа, 2019",
-      title: "Национальное достояние – парки",
-      paragraph: "В 2016 году Америка отмечала важный юбилей: сто лет назад здесь начала складываться система национальных парков – охраняемых территорий, где и сегодня каждый может приобщиться к природе.",
-      subtitle: "Дзен",
-    },
-    {
-      keyWord: "Новая Ладога длинное слово",
-      date: "2 августа, 2019",
-      title: "Национальное достояние – парки",
-      paragraph: "В 2016 году Америка отмечала важный юбилей: сто лет назад здесь начала складываться система национальных парков – охраняемых территорий, где и сегодня каждый может приобщиться к природе.",
-      subtitle: "Дзен",
-    },
-    {
-      keyWord: "Ладога",
-      date: "2 августа, 2019",
-      title: "Национальное достояние – парки",
-      paragraph: "В 2016 году Америка отмечала важный юбилей: сто лет назад здесь начала складываться система национальных парков – охраняемых территорий, где и сегодня каждый может приобщиться к природе.",
-      subtitle: "Дзен",
-    },
-    {
-      keyWord: "Ладога",
-      date: "2 августа, 2019",
-      title: "Национальное достояние – парки",
-      paragraph: "В 2016 году Америка отмечала важный юбилей: сто лет назад здесь начала складываться система национальных парков – охраняемых территорий, где и сегодня каждый может приобщиться к природе.",
-      subtitle: "Дзен",
-    },
-    {
-      keyWord: "Ладога",
-      date: "2 августа, 2019",
-      title: "Национальное достояние – парки",
-      paragraph: "В 2016 году Америка отмечала важный юбилей: сто лет назад здесь начала складываться система национальных парков – охраняемых территорий, где и сегодня каждый может приобщиться к природе.",
-      subtitle: "Дзен",
-    }
-  ]
-
-  const user = { name: "Стасон" }
 
   const [isLoginPopupOpen, setIsLoginPopupOpen] = useState();
   const [HeaderButtonName, setHeaderButtonName] = useState('Авторизоваться');
