+++ src/utils/constants.js
@@ -1,0 +1,68 @@
+const user = { name: "Стасон" }
+const cards = [
+    {
+        keyWord: "Ладога",
+        date: "2 августа, 2019",
+        title: "Национальное достояние – парки",
+        paragraph: "В 2016 году Америка отмечала важный юбилей: сто лет назад здесь начала складываться система национальных парков – охраняемых территорий, где и сегодня каждый может приобщиться к природе.",
+        subtitle: "Дзен",
+    },
+    {
+        keyWord: "Ладога",
+        date: "2 августа, 2019",
+        title: "Национальное достояние – парки",
+        paragraph: "В 2016 году Америка отмечала важный юбилей: сто лет назад здесь начала складываться система национальных парков – охраняемых территорий, где и сегодня каждый может приобщиться к природе.",
+        subtitle: "Дзен",
+    },
+    {
+        keyWord: "Ладога",
+        date: "2 августа, 2019",
+        title: "Национальное достояние – парки",
+        paragraph: "В 2016 году Америка отмечала важный юбилей: сто лет назад здесь начала складываться система национальных парков – охраняемых территорий, где и сегодня каждый может приобщиться к природе.",
+        subtitle: "Дзен",
+    },
+    {
+        keyWord: "Ладога",
+        date: "2 августа, 2019",
+        title: "Национальное достояние – парки",
+        paragraph: "В 2016 году Америка отмечала важный юбилей: сто лет назад здесь начала складываться система национальных парков – охраняемых территорий, где и сегодня каждый может приобщиться к природе.",
+        subtitle: "Дзен",
+    },
+    {
+        keyWord: "Ладога",
+        date: "2 августа, 2019",
+        title: "Национальное достояние – парки",
+        paragraph: "В 2016 году Америка отмечала важный юбилей: сто лет назад здесь начала складываться система национальных парков – охраняемых территорий, где и сегодня каждый может приобщиться к природе.",
+        subtitle: "Дзен",
+    },
+    {
+        keyWord: "Новая Ладога длинное слово",
+        date: "2 августа, 2019",
+        title: "Национальное достояние – парки",
+        paragraph: "В 2016 году Америка отмечала важный юбилей: сто лет назад здесь начала складываться система национальных парков – охраняемых территорий, где и сегодня каждый может приобщиться к природе.",
+        subtitle: "Дзен",
+    },
+    {
+        keyWord: "Ладога",
+        date: "2 августа, 2019",
+        title: "Национальное достояние – парки",
+        paragraph: "В 2016 году Америка отмечала важный юбилей: сто лет назад здесь начала складываться система национальных парков – охраняемых территорий, где и сегодня каждый может приобщиться к природе.",
+        subtitle: "Дзен",
+    },
+    {
+        keyWord: "Ладога",
+        date: "2 августа, 2019",
+        title: "Национальное достояние – парки",
+        paragraph: "В 2016 году Америка отмечала важный юбилей: сто лет назад здесь начала складываться система национальных парков – охраняемых территорий, где и сегодня каждый может приобщиться к природе.",
+        subtitle: "Дзен",
+    },
+    {
+        keyWord: "Ладога",
+        date: "2 августа, 2019",
+        title: "Национальное достояние – парки",
+        paragraph: "В 2016 году Америка отмечала важный юбилей: сто лет назад здесь начала складываться система национальных парков – охраняемых территорий, где и сегодня каждый может приобщиться к природе.",
+        subtitle: "Дзен",
+    }
+]
+
+export { cards, user };
